re-organize code to be able to define Show,Eq,Ord,Data,Monoid,IsList without orphans
