+++ Foundation/Primitive/Block.hs
@@ -27,16 +27,9 @@ module Foundation.Primitive.Block
     , thaw
     , copy
     -- * safer api
-    , empty
     , create
     , singleton
     , replicate
-    , fromList
-    , toList
-    , equal
-    , bcompare
-    , append
-    , concat
     , index
     , map
     , foldl
@@ -50,15 +43,16 @@ module Foundation.Primitive.Block
     , splitAt
     , break
     , span
+    , elem
+    , all
+    , any
     -- * Foreign interfaces
     , unsafeCopyToPtr
     ) where
 
 import           GHC.Prim
 import           GHC.Types
---import           GHC.Ptr
 import           GHC.ST
-import qualified Data.List
 import           Foundation.Internal.Base hiding (fromList, toList)
 import           Foundation.Internal.Proxy
 import           Foundation.Internal.Primitive
@@ -69,41 +63,13 @@ import           Foundation.Primitive.IntegralConv
 import           Foundation.Primitive.Types
 import qualified Foundation.Primitive.Block.Mutable as M
 import           Foundation.Primitive.Block.Mutable (Block(..), MutableBlock(..), new, unsafeThaw, unsafeFreeze)
+import           Foundation.Primitive.Block.Base
 import           Foundation.Numerical
-
--- | Create an empty block of memory
-empty :: Block ty
-empty = Block ba where !(Block ba) = empty_
-
-empty_ :: Block ()
-empty_ = runST $ primitive $ \s1 ->
-    case newByteArray# 0# s1           of { (# s2, mba #) ->
-    case unsafeFreezeByteArray# mba s2 of { (# s3, ba  #) ->
-        (# s3, Block ba #) }}
 
 -- | return the number of elements of the array.
 length :: PrimType ty => Block ty -> Int
 length a = let (Size len) = lengthSize a in len
 {-# INLINE[1] length #-}
-
-lengthSize :: forall ty . PrimType ty => Block ty -> Size ty
-lengthSize (Block ba) =
-    let !(Size (I# szBits)) = primSizeInBytes (Proxy :: Proxy ty)
-        !elems              = quotInt# (sizeofByteArray# ba) szBits
-     in Size (I# elems)
-{-# INLINE[1] lengthSize #-}
-
-lengthBytes :: Block ty -> Size Word8
-lengthBytes (Block ba) = Size (I# (sizeofByteArray# ba))
-{-# INLINE[1] lengthBytes #-}
-
--- | Return the element at a specific index from an array without bounds checking.
---
--- Reading from invalid memory can return unpredictable and invalid values.
--- use 'index' if unsure.
-unsafeIndex :: forall ty . PrimType ty => Block ty -> Offset ty -> ty
-unsafeIndex (Block ba) n = primBaIndex ba n
-{-# INLINE unsafeIndex #-}
 
 -- | Copy all the block content to the memory starting at the destination address
 unsafeCopyToPtr :: forall ty prim . PrimMonad prim
@@ -120,7 +86,7 @@ create :: forall ty . PrimType ty
        -> (Offset ty -> ty) -- ^ the function that set the value at the index
        -> Block ty          -- ^ the array created
 create n initializer
-    | n == 0    = empty
+    | n == 0    = mempty
     | otherwise = runST $ do
         mb <- new n
         M.iterSet initializer mb
@@ -131,89 +97,6 @@ singleton ty = create 1 (const ty)
 
 replicate :: PrimType ty => Word -> ty -> Block ty
 replicate sz ty = create (Size (integralCast sz)) (const ty)
-
--- | make a block from a list of elements.
-fromList :: PrimType ty => [ty] -> Block ty
-fromList l = runST $ do
-    ma <- new (Size len)
-    iter 0 l $ \i x -> M.unsafeWrite ma i x
-    unsafeFreeze ma
-  where len = Data.List.length l
-        iter _  []     _ = return ()
-        iter !i (x:xs) z = z i x >> iter (i+1) xs z
-
--- | transform a block to a list.
-toList :: forall ty . PrimType ty => Block ty -> [ty]
-toList blk@(Block ba)
-    | len == 0  = []
-    | otherwise = loop 0
-  where
-    !len = lengthSize blk
-    loop !i | i .==# len = []
-            | otherwise  = primBaIndex ba i : loop (i+1)
-
--- | Check if two vectors are identical
-equal :: (PrimType ty, Eq ty) => Block ty -> Block ty -> Bool
-equal a b
-    | la /= lb  = False
-    | otherwise = loop 0
-  where
-    !la = lengthSize a
-    !lb = lengthSize b
-    loop n | n .==# la = True
-           | otherwise = (unsafeIndex a n == unsafeIndex b n) && loop (n+1)
-
--- | Compare 2 vectors
-bcompare :: (Ord ty, PrimType ty) => Block ty -> Block ty -> Ordering
-bcompare a b = loop 0
-  where
-    !la = lengthSize a
-    !lb = lengthSize b
-    loop n
-        | n .==# la = if la == lb then EQ else LT
-        | n .==# lb = GT
-        | otherwise =
-            case unsafeIndex a n `compare` unsafeIndex b n of
-                EQ -> loop (n+1)
-                r  -> r
-
--- | Append 2 arrays together by creating a new bigger array
-append :: Block ty -> Block ty -> Block ty
-append a b
-    | la == 0 = b
-    | lb == 0 = a
-    | otherwise = runST $ do
-        r  <- M.unsafeNew (la+lb)
-        M.unsafeCopyBytesRO r 0                 a 0 la
-        M.unsafeCopyBytesRO r (sizeAsOffset la) b 0 lb
-        unsafeFreeze r
-  where
-    !la = lengthBytes a
-    !lb = lengthBytes b
-
-concat :: [Block ty] -> Block ty
-concat [] = empty
-concat l  =
-    case filterAndSum 0 [] l of
-        (_,[])            -> empty
-        (_,[x])           -> x
-        (totalLen,chunks) -> runST $ do
-            r <- M.unsafeNew totalLen
-            doCopy r 0 chunks
-            unsafeFreeze r
-  where
-    -- TODO would go faster not to reverse but pack from the end instead
-    filterAndSum !totalLen acc []     = (totalLen, Data.List.reverse acc)
-    filterAndSum !totalLen acc (x:xs)
-        | len == 0  = filterAndSum totalLen acc xs
-        | otherwise = filterAndSum (len+totalLen) (x:acc) xs
-      where len = lengthBytes x
-
-    doCopy _ _ []     = return ()
-    doCopy r i (x:xs) = do
-        M.unsafeCopyBytesRO r i x 0 lx
-        doCopy r (i `offsetPlusE` lx) xs
-      where !lx = lengthBytes x
 
 -- | Thaw a Block into a MutableBlock
 --
@@ -294,7 +177,7 @@ snoc vec e
 
 sub :: PrimType ty => Block ty -> Offset ty -> Offset ty -> Block ty
 sub vec start end
-    | start >= end = empty
+    | start >= end = mempty
     | otherwise    = runST $ do
         dst <- new len
         M.unsafeCopyElementsRO dst 0 vec start newLen
@@ -321,8 +204,8 @@ unsnoc vec
 
 splitAt :: PrimType ty => Size ty -> Block ty -> (Block ty, Block ty)
 splitAt nbElems v
-    | nbElems <= 0 = (empty, v)
-    | n == vlen    = (v, empty)
+    | nbElems <= 0 = (mempty, v)
+    | n == vlen    = (v, mempty)
     | otherwise    = runST $ do
         left  <- new nbElems
         right <- new (vlen - nbElems)
@@ -339,10 +222,37 @@ break predicate blk = findBreak 0
   where
     !len = lengthSize blk
     findBreak !i
-        | i .==# len                    = (blk, empty)
+        | i .==# len                    = (blk, mempty)
         | predicate (unsafeIndex blk i) = splitAt (offsetAsSize i) blk
         | otherwise                     = findBreak (i + 1)
     {-# INLINE findBreak #-}
 
 span :: PrimType ty => (ty -> Bool) -> Block ty -> (Block ty, Block ty)
 span p = break (not . p)
+
+elem :: PrimType ty => ty -> Block ty -> Bool
+elem v blk = loop 0
+  where
+    !len = lengthSize blk
+    loop i
+        | i .==# len             = False
+        | unsafeIndex blk i == v = True
+        | otherwise              = loop (i+1)
+
+all :: PrimType ty => (ty -> Bool) -> Block ty -> Bool
+all p blk = loop 0
+  where
+    !len = lengthSize blk
+    loop i
+        | i .==# len            = True
+        | p (unsafeIndex blk i) = loop (i+1)
+        | otherwise             = False
+
+any :: PrimType ty => (ty -> Bool) -> Block ty -> Bool
+any p blk = loop 0
+  where
+    !len = lengthSize blk
+    loop i
+        | i .==# len            = False
+        | p (unsafeIndex blk i) = True
+        | otherwise             = loop (i+1)
